pyproject.toml
[build-system]
requires = [
  "scikit-build-core>=0.9",
  "setuptools>=68",
  "wheel",
  "cmake>=3.22",
  "ninja",
  "pybind11>=2.6.0",
  "conan>=2.0.0"
]
build-backend = "scikit_build_core.build"

[project]
name = "pulseflow"
version = "0.1.1"
description = "High-performance radio astronomy single pulse Detect lib"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "BSD" }
authors = [{ name = "lintian233" }]

dependencies = [
  "astropy==7.0.1",
  "loguru==0.7.3",
  "matplotlib==3.10.5",
  "numba==0.61.0",
  "numpy==2.1.3",
  "opencv_python==4.11.0.86",
  "pandas==2.3.1",
  "PyYAML==6.0.2",
  "scipy==1.16.1",
  "seaborn==0.13.2",
  "sympy==1.13.1",
  "torch==2.6.0",
  "torchvision==0.21.0",
  "tqdm==4.67.1",
  "typing_extensions==4.14.1",
  "ultralytics==8.3.167",
  "your==0.6.7"
]

[project.optional-dependencies]
dev = ["pytest>=6.0", "pytest-cov", "black", "flake8"]

[project.scripts]
astroflow = "astroflow.cli:main"
astrorfimask = "astroflow.rfimask:astrorfimask"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"" = "python"

[tool.setuptools.packages.find]
where = ["python"]

[tool.setuptools.package-data]
astroflow = ["lib/*/*.{so,dll,dylib}"]

[tool.scikit-build]
wheel.packages = ["python/astroflow"]

[tool.scikit-build.cmake]
minimum-version = "3.22"

[tool.scikit-build.cmake.define]
CMAKE_BUILD_TYPE = "Release"

[tool.scikit-build.sdist]
include = [
  "CMakeLists.txt",
  "conanfile.txt",
  "include/**",
  "src/**",
  "python/**",
]
exclude = [
  "build/**",
  "_skbuild/**",
  "dist/**",
  ".git/**",
  ".github/**",
]

[tool.cibuildwheel]
build = ["cp310-*", "cp311-*", "cp312-*"]
archs = ["x86_64"]
skip = ["*-musllinux_*", "pp*"]

[tool.cibuildwheel.linux]
repair-wheel-command = "auditwheel repair -w {dest_dir} {wheel}"


[tool.cibuildwheel.environment]
CMAKE_BUILD_TYPE = "Release"
CMAKE_CUDA_ARCHITECTURES = "75;80;86;89"
CUDAHOME = "/usr/local/cuda"
CUDACXX = "/usr/local/cuda/bin/nvcc"
NVCC_PREPEND_FLAGS = "--allow-unsupported-compiler"